mxtaltools/standalone/autoencoder/Mo3ENet/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Mo3ENet"
version = "0.0.1"
authors = [
  { name="Michael Kilgour", email="mjakilgour@gmail.com" },
]
description = "Package containing a model for the prediction of the packing coefficient for a molecular crystal, given single-molecule information."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "pytorch",
    "pytorch-geometric",
    "torch_scatter",
    "torch_cluster",
    "yaml",
    "plotly",
]
[tool.hatch.build.targets.sdist]
include = [
    "Mo3ENet/*"
]
[project.urls]
Homepage = "https://github.com/InfluenceFunctional/Mo3ENet"
Issues = "https://github.com/InfluenceFunctional/Mo3ENet/issues"